+++ ast/src/input/definition.rs
@@ -14,29 +14,13 @@
 // You should have received a copy of the GNU General Public License
 // along with the Leo library. If not, see <https://www.gnu.org/licenses/>.
 
-pub mod definition;
-pub use definition::*;
+use super::*;
+use crate::{Expression, Identifier, Type};
 
-pub mod input;
-pub use input::*;
-
-pub mod input_value;
-pub use input_value::*;
-
-pub mod parameter;
-pub use parameter::*;
-
-pub mod program_input;
-pub use program_input::*;
-
-pub mod program_state;
-pub use program_state::*;
-
-pub mod section;
-pub use section::*;
-
-use indexmap::IndexMap;
-use leo_errors::{LeoError, Result};
-use leo_span::{sym, Span, Symbol};
-
-type Definitions = IndexMap<Parameter, InputValue>;
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct Definition {
+    pub type_: Type,
+    pub name: Identifier,
+    pub value: Expression,
+    pub span: Span,
+}
