+++ ast/src/input/parameter.rs
@@ -14,29 +14,24 @@
 // You should have received a copy of the GNU General Public License
 // along with the Leo library. If not, see <https://www.gnu.org/licenses/>.
 
-pub mod definition;
-pub use definition::*;
+use super::*;
+use crate::{Identifier, Type};
 
-pub mod input;
-pub use input::*;
+/// A set of properties for a single definition in an input file.
+/// Used as a key in [`ProgramInput`] and [`ProgramState`].
+#[derive(Debug, Clone, PartialEq, Eq, Hash)]
+pub struct Parameter {
+    pub variable: Identifier,
+    pub type_: Type,
+    pub span: Span,
+}
 
-pub mod input_value;
-pub use input_value::*;
-
-pub mod parameter;
-pub use parameter::*;
-
-pub mod program_input;
-pub use program_input::*;
-
-pub mod program_state;
-pub use program_state::*;
-
-pub mod section;
-pub use section::*;
-
-use indexmap::IndexMap;
-use leo_errors::{LeoError, Result};
-use leo_span::{sym, Span, Symbol};
-
-type Definitions = IndexMap<Parameter, InputValue>;
+impl From<Definition> for Parameter {
+    fn from(definition: Definition) -> Self {
+        Self {
+            variable: definition.name,
+            type_: definition.type_,
+            span: definition.span,
+        }
+    }
+}
