+++ parser/src/parser/input.rs
@@ -14,7 +14,6 @@
 // You should have received a copy of the GNU General Public License
 // along with the Leo library. If not, see <https://www.gnu.org/licenses/>.
 
-
 use super::*;
 
 use leo_errors::{ParserError, Result};
